separate post rides test
