+++ src/app.js
@@ -19,37 +19,37 @@ module.exports = (db) => {
         const driverVehicle = req.body.driver_vehicle;
 
         if (startLatitude < -90 || startLatitude > 90 || startLongitude < -180 || startLongitude > 180) {
-            return res.send({
+            return res.status(400).send({
                 error_code: 'VALIDATION_ERROR',
                 message: 'Start latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
             });
         }
 
         if (endLatitude < -90 || endLatitude > 90 || endLongitude < -180 || endLongitude > 180) {
-            return res.send({
+            return res.status(400).send({
                 error_code: 'VALIDATION_ERROR',
                 message: 'End latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
             });
         }
 
         if (typeof riderName !== 'string' || riderName.length < 1) {
-            return res.send({
+            return res.status(400).send({
                 error_code: 'VALIDATION_ERROR',
                 message: 'Rider name must be a non empty string'
             });
         }
 
         if (typeof driverName !== 'string' || driverName.length < 1) {
-            return res.send({
+            return res.status(400).send({
                 error_code: 'VALIDATION_ERROR',
-                message: 'Rider name must be a non empty string'
+                message: 'Driver name must be a non empty string'
             });
         }
 
         if (typeof driverVehicle !== 'string' || driverVehicle.length < 1) {
-            return res.send({
+            return res.status(400).send({
                 error_code: 'VALIDATION_ERROR',
-                message: 'Rider name must be a non empty string'
+                message: 'Driver vehicle must be a non empty string'
             });
         }
 
@@ -57,7 +57,7 @@ module.exports = (db) => {
         
         db.run('INSERT INTO Rides(startLat, startLong, endLat, endLong, riderName, driverName, driverVehicle) VALUES (?, ?, ?, ?, ?, ?, ?)', values, function (err) {
             if (err) {
-                return res.send({
+                return res.status(400).send({
                     error_code: 'SERVER_ERROR',
                     message: 'Unknown error'
                 });
@@ -86,7 +86,7 @@ module.exports = (db) => {
             }
 
             if (rows.length === 0) {
-                return res.send({
+                return res.status(400).send({
                     error_code: 'RIDES_NOT_FOUND_ERROR',
                     message: 'Could not find any rides'
                 });
@@ -106,7 +106,7 @@ module.exports = (db) => {
             }
 
             if (rows.length === 0) {
-                return res.send({
+                return res.status(400).send({
                     error_code: 'RIDES_NOT_FOUND_ERROR',
                     message: 'Could not find any rides'
                 });
